@@ -1,6 +1,5 @@
-let allNotesRu = ['ДО', 'РЕ', 'МИ', 'ФА', 'СОЛЬ', 'ЛЯ', 'СИ'];
-let allNotesEn = ['C', 'D', 'E', 'F', 'G', 'A', 'H'];
-
-let allNumbers = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9'];
+const allNotesRu = ['ДО', 'РЕ', 'МИ', 'ФА', 'СОЛЬ', 'ЛЯ', 'СИ'];
+const allNotesEn = ['C', 'D', 'E', 'F', 'G', 'A', 'H'];
+const allNumbers = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9'];
 
 export { allNotesRu, allNotesEn, allNumbers };
