@@ -1,5 +1,5 @@
 import { ITask } from './ITask';
-export type operation = 'plus' | 'minus';
+export type operation = 'plus' | 'minus' | 'multiply' | 'divide';
 
 export class MathProblem implements ITask {
   a1: number;
@@ -11,9 +11,12 @@ export class MathProblem implements ITask {
     this.operation = operation;
   }
   solveProblem = (): number =>
-    this.operation === 'plus' ? this.a1 + this.a2 : this.a1 - this.a2;
+    this.operation === 'plus' ? this.a1 + this.a2 
+      : this.operation === 'minus' ? this.a1 - this.a2
+      : this.operation === 'multiply' ? this.a1 * this.a2
+      : this.a1 / this.a2;
   getProblemComplexity = () => {
-    return (this.a1 + this.a2) / 50;
+    return (this.a1 + this.a2) / 100;
   };
 
   isCorrectAnswer = (answer: number) => answer === this.solveProblem();
@@ -29,14 +32,23 @@ const randomInt = (min: number, max: number): number => {
 };
 
 export const generateMathProblem = (): MathProblem => {
-  const a1 = randomInt(0, 29);
-  const a2 = randomInt(0, 19);
-  const op: operation = randomInt(0, 1) === 0 ? 'plus' : 'minus';
+  const opCode = randomInt(0, 3);
+  const op: operation = opCode === 0 ? 'plus' : opCode === 1 ? 'minus': opCode === 2 ? 'multiply' : 'divide';
+  let a1 = randomInt(1, 49);
+  let a2 = randomInt(1, 39);
+  if(op == 'multiply') {
+    a1 = randomInt(1, 9);
+    a2 = randomInt(1, 9);
+  }
+  if(op == 'divide') {
+    a1 = randomInt(1, 9);
+    a2 = randomInt(1, 9);
+    a1 = a1 * a2;
+  }
 
-  var problem: MathProblem = new MathProblem(a1, a2, op);
+  const problem: MathProblem = new MathProblem(a1, a2, op);
 
   if (problem.solveProblem() < 0) return generateMathProblem();
-  if (a1 === 0 || a2 === 0) return generateMathProblem();
-
+  
   return problem;
 };
